Small fixups to `ConvOp` and some restructuring, renaming
Note: This also clea `ConvOp` using `Number` instead of `Int`, since
it might in general take either float or int.

It also removes the `trap`ing of `i64.wrap_i64`, since it is a non-existent
instruction, but trapping is something an actual program might do.
Better for the interpreter to get stuck.